pyproject.toml
[tool.poetry]
name = "parrot-tools"
version = "0.1.0"
description = "A collection of tools for text2image"
authors = ["Stephen Young"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
diffusers = "^0.2.4"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.8.0"
flake8 = "^5.0.4"
mypy = "^0.971"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
rich = "^12.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
